# src/main/java/com/example/bankcards/dto/request/AuthUserRequest.java
package com.example.bankcards.dto.request;


import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
@Schema(description = "Запрос на аутентификацию")
public record AuthUserRequest(

        @Schema(description = "Имя пользователя", example = "username")
        @NotBlank(message = "Имя пользователя обязательно")
        @Size(min = 4, max = 50, message = "Имя пользователя должно быть от 4 до 50 символов")
        String username,
        @Schema(description = "Пароль", example = "password")
        @NotBlank(message = "Пароль обязателен")
        @Size(min = 4, max = 100, message = "Пароль должен быть от 4 до 100 символов")
        String password) {

}